+++ build_wheel.sh
@@ -1,0 +1,3 @@
+#! bash
+pip install build
+python -m build --wheel --outdir dist